requirements.txt
requests==2.31.0
SQLAlchemy==2.0.23
aiogram==3.3.0
schedule==1.2.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
psutil==7.1.0
pytz==2023.3